requirements.txt
pandas
pyarrow
pillow
matplotlib 